third_party/blink/web_tests/platform/mac-mac12-arm64/external/wpt/html/cross-origin-opener-policy/reporting/navigation-reporting/reporting-redirect-with-same-origin-allow-popups.https-expected.txt

This is a testharness.js-based test.
FAIL Same origin openee redirected to same-origin with same-origin-allow-popups assert_equals: expected (string) "coop" but got (undefined) undefined
PASS Cross origin openee redirected to same-origin with same-origin-allow-popups
Harness: the test ran to completion.
